Remise en conformité écran Admin et simplification
